% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_names}
\alias{cat_names}
\title{Names of awesome cats}
\format{Atomic character vector with 3 elements:
\itemize{
\item Walter: White, deaf and awesome. Likes chemistry.
\item Mollie: Super smart and kind. Likes nature.
\item Dexter: Big and ginger. Likes bouncy balls.
}}
\usage{
cat_names
}
\description{
A vector containing names of cats.
Some are cereal killers.
}
\keyword{datasets}
