% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covis-package.R
\docType{package}
\name{covis-package}
\alias{covis-package}
\title{covis: Coverage Visualisation for Cancer Gene Panels}
\description{
The \code{covis} package provides functions for plotting
coverage depth across cancer gene panels.
}
\details{
\itemize{
\item Foo
\item Bar
\item Baz
}
}
\section{Foo functions}{

The foo functions do this, this and that.
}

