% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.R
\name{read_bed}
\alias{read_bed}
\title{Read a BED file}
\usage{
read_bed(fn, param = NULL)
}
\arguments{
\item{fn}{BED file name. Must be a character vector of length 1.}

\item{param}{A GRanges or GRangesList instance, used to select
which records to scan.}
}
\value{
A GRanges object representation of the input BED file.
}
\description{
Read a BED file output by mosdepth, containing four
columns: chrom, start_pos, end_pos and coverage. If a GRanges object
is specified for \code{param}, it scans only those records from the BED file.
}
\details{
The BED file needs to be bgzipped and tbi-indexed for use with tabix.
\strong{Note} that a csi-index (mosdepth default) is not supported
by Rsamtools (or maybe it just doesn't work for  me).

For example, you can run the following on a sorted BED file:
\code{bgzip <sample>.per-base.bed; tabix -p bed <sample>.per-base.bed.gz},
and then read the resulting gzipped and indexed file with this function.
}
\examples{
fn <- system.file("extdata", "mosdepth_s1.per-base.bed.gz", package = "covis")
gr <- GenomicRanges::GRanges(c("1:10000-12000", "1:13000-14000"))
x1 <- read_bed(fn)
x2 <- read_bed(fn, gr)

}
